CF rocket loader turnstile script not deffered
Hi,
Using your plugin I came to issue of a website that if Cloudflare Rocket loader is ON, Turnsile box wasn't displayed because one of Turnsite scripts was loaded too late.
In order to make it load with Rocket loader ON is to add attribute data-cfasync="false" and that will exempt the script from differing.
I've tested this on few sites and works well. Please include it, I think more people will benefit. 
Best regards,
Jaro